--- packages/stack-supabase/src/query-builder-v3.ts
@@ -1,989 +1,0 @@
-import {
-  DATE_LIKE_CASTS,
-  EncryptedV3Column,
-  logger,
-  matchNeedleError,
-  parseSelectorSegments,
-  reconstructSelectorDocument,
-  unsupportedLeafReason,
-} from '@cipherstash/stack/adapter-kit'
-import type { EncryptionClient } from '@cipherstash/stack/encryption'
-import type { AnyV3Table } from '@cipherstash/stack/eql/v3'
-import {
-  type EncryptionError,
-  EncryptionErrorTypes,
-} from '@cipherstash/stack/errors'
-import type {
-  ColumnSchema,
-  EncryptedTable,
-  EncryptedTableColumn,
-} from '@cipherstash/stack/schema'
-import type {
-  BuildableQueryColumn,
-  Encrypted,
-  EncryptedQueryResult,
-  QueryTypeName,
-  ScalarQueryTerm,
-} from '@cipherstash/stack/types'
-import { addJsonbCastsV3, selectKeyToDbV3 } from './helpers'
-import {
-  EncryptedQueryBuilderImpl,
-  EncryptionFailedError,
-} from './query-builder'
-import type {
-  DbName,
-  DbSelect,
-  FilterOp,
-  SupabaseClientLike,
-  SupabaseQueryBuilder,
-} from './types'
-
-/** cast_as kinds that reconstruct to a JS `Date` — shared with the typed v3
- * client's decrypt-model path (see `encryption/v3.ts`). */
-const DATE_LIKE_CAST_SET = new Set<string>(DATE_LIKE_CASTS)
-
-/**
- * The subset of a v3 column builder the dialect relies on. Structural rather
- * than the concrete class union so the runtime `instanceof EncryptedV3Column`
- * gate and this type stay independent.
- */
-type V3ColumnLike = {
-  getName(): string
-  getEqlType(): string
-  getQueryCapabilities(): {
-    equality: boolean
-    orderAndRange: boolean
-    freeTextSearch: boolean
-    /** Optional: only `public.eql_v3_json_search` (`types.Json`) carries it. */
-    searchableJson?: boolean
-  }
-  build(): ColumnSchema
-}
-
-/**
- * Validate an encrypted-JSON containment operand: a NON-EMPTY plain object or a
- * non-empty array. Everything else is rejected with an actionable steer:
- *
- * - Scalars/strings: the caller meant free-text (`matches` on a text column) or
- *   a selector — a raw JSON string is NOT parsed, by design (parsing would make
- *   `'{"a":1}'` and `{a:1}` silently different queries on other surfaces).
- * - Non-plain objects (`Date`, `Map`, `RegExp`, class instances): these JSON-
- *   serialize to scalars or `{}` — not the sub-document the caller believes.
- * - `{}` and `[]`: jsonb containment holds for EVERY document (`doc @> '{}'`),
- *   so an accidentally-empty needle would silently return (and decrypt) the
- *   whole table. The Drizzle adapter rejects the same needle for the same
- *   reason — the two first-party adapters must agree that this is an error.
- */
-function assertJsonContainmentOperand(column: string, value: unknown): void {
-  const isPlainObject =
-    value !== null &&
-    typeof value === 'object' &&
-    !Array.isArray(value) &&
-    (Object.getPrototypeOf(value) === Object.prototype ||
-      Object.getPrototypeOf(value) === null)
-  if (!isPlainObject && !Array.isArray(value)) {
-    // Array.isArray is false on this branch by construction, so the label only
-    // distinguishes null / non-plain object / scalar.
-    const got =
-      value === null
-        ? 'null'
-        : typeof value === 'object'
-          ? (value as object).constructor?.name || 'a non-plain object'
-          : typeof value
-    throw new Error(
-      `[supabase v3]: encrypted JSON containment on column "${column}" takes a sub-document (plain object or array) to match, got ${got}.`,
-    )
-  }
-  const empty = Array.isArray(value)
-    ? value.length === 0
-    : Object.keys(value as object).length === 0
-  if (empty) {
-    throw new Error(
-      `[supabase v3]: encrypted JSON containment on column "${column}" cannot take an empty ${Array.isArray(value) ? 'array' : 'object'} needle: it matches every row. Pass a non-empty sub-document, or omit the predicate to select all rows.`,
-    )
-  }
-}
-
-/**
- * Reject a declared property name that is also a DIFFERENT physical column.
- *
- * `select('*')` expands the introspected DB names into property names, so a
- * column renamed `created_at → createdAt` and a distinct plaintext column
- * literally named `createdAt` both emit the token `createdAt`, which
- * `addJsonbCastsV3` turns into `createdAt:created_at::jsonb` — twice. PostgREST
- * returns the encrypted column under that key and the plaintext one is never
- * selected, silently yielding the wrong value for a field the row type
- * guarantees.
- *
- * Nothing downstream can disambiguate the two, and `EncryptedTable.build()`'s
- * duplicate check only fires when two BUILDERS share a `getName()`. Refuse to
- * construct instead.
- */
-function assertNoPropertyDbNameCollision(
-  tableName: string,
-  propToDb: Record<string, string>,
-  allColumns: string[] | null,
-): void {
-  if (!allColumns) return
-  const dbNames = new Set(allColumns)
-
-  for (const [property, dbName] of Object.entries(propToDb)) {
-    if (property === dbName) continue
-    if (!dbNames.has(property)) continue
-    throw new Error(
-      `[supabase v3]: property "${property}" on table "${tableName}" renames DB column "${dbName}", but "${property}" is also a distinct column in the database — the two collide in select('*'). Rename the property, or drop the declared rename.`,
-    )
-  }
-}
-
-/**
- * EQL v3 dialect of {@link EncryptedQueryBuilderImpl} for native concrete-domain
- * columns (`public.*` type domains, `eql_v3` operators). The query mechanism is
- * v2's — direct EQL operators over PostgREST — with four narrow forks:
- *
- * - **Column recognition / naming** — v3 columns are `EncryptedV3Column`
- *   builders and may map a JS property name to a different DB column name
- *   (`buildColumnKeyMap`). Filters, select casts, and mutations resolve
- *   property → DB name; select casts alias the DB column back to the property
- *   (`prop:db_name::jsonb`) so result rows keep property keys.
- * - **Mutation encoding** — the raw encrypted payload object is sent (the
- *   `public.*` domains are `DOMAIN … AS jsonb`), not v2's `{ data: … }`
- *   composite wrap.
- * - **Query-term encoding** — scalar equality/range filters use the FULL
- *   storage envelope from `encrypt()`, serialized as jsonb text.
- *
- *   NOT because narrowed terms fail the domain CHECK: the bundle defines a
- *   `public.<domain>_query` companion for each storage domain, whose CHECK
- *   requires `NOT (VALUE ? 'c')` — i.e. it accepts exactly the no-ciphertext
- *   shape `encryptQuery` produces. Those domains are simply unreachable from
- *   here. PostgREST has no syntax to cast a filter VALUE, and an uncast literal
- *   is ambiguous between the `_query` and `jsonb` `@>`/`=` overloads (42725 —
- *   the bundle says so itself, see `cipherstash-encrypt-v3.sql`, the
- *   `_query_types.sql` note). The reachable overload is the `jsonb` one, whose
- *   body coerces its operand to the STORAGE domain, which does require `c`.
- *   (protect-ffi can mint narrowed `eql_v3.query_<name>` operands via
- *   `encryptQuery`, but with no way to cast a PostgREST filter value they
- *   stay unreachable from this adapter.)
- *
- *   The full envelope satisfies scalar storage-domain CHECKs by construction,
- *   and equality/range operators extract the term they need.
- *
- *   EQL 3.0.2 removed the storage/jsonb escape hatch for free-text and JSON
- *   operators: those now require typed query-domain operands. The factory reads
- *   the installed EQL version and this builder fails those operators before
- *   encryption, so a decryptable storage envelope never enters a GET URL.
- * - **Legacy `matches`, not `like`/`ilike`/`contains`** — on pre-3.0.2 EQL,
- *   encrypted free-text search is
- *   FUZZY BLOOM TOKEN MATCHING, not containment: the bundle declares `@>` on each
- *   match domain (`CREATE OPERATOR @> … FUNCTION = eql_v3.matches`, the SQL
- *   function's name), whose body is `match_term(a) @> match_term(b)` — `smallint[]`
- *   containment of the two bloom filters. It is order- and multiplicity-
- *   insensitive and one-sided (a `true` may be a false positive). PostgREST
- *   reaches it as `cs`. The operator is named `matches` to signal that; `contains`
- *   is reserved for exact (native) containment on plaintext columns.
- *
- *   Match is tokenized + downcased, so `%` is NOT a wildcard. `like`/`ilike` on an
- *   encrypted column are delegated to `matches` as an APPROXIMATE compatibility
- *   shim (surrounding `%` stripped, internal `%`/`_` rejected, one warning) and
- *   pass through as real SQL LIKE on a plaintext column.
- *
- *   Substrings DO match: the needle blooms to its own trigrams, and containment
- *   holds whenever every one of them is present in the stored value's bloom —
- *   i.e. for any substring of at least `token_length` (3) characters. Shorter
- *   needles bloom to nothing (`bf @> '{}'` is true for every row) and are
- *   rejected up front by `matchNeedleError`, not answered.
- *
- * Decrypted rows additionally get `Date` reconstruction from the
- * encrypt-config `cast_as`, mirroring the typed v3 client.
- */
-export class EncryptedQueryBuilderV3Impl<
-  T extends Record<string, unknown> = Record<string, unknown>,
-> extends EncryptedQueryBuilderImpl<T> {
-  private v3Table: AnyV3Table
-  /** JS property name → DB column name, for every encrypted column. */
-  private propToDb: Record<string, string>
-  /** DB column name → JS property name — the inverse of {@link propToDb}, used
-   * to expand `select('*')` back into property names. Null prototype: a DB
-   * column literally named `constructor` / `toString` would otherwise resolve
-   * to an inherited `Object.prototype` member and be emitted as a select token. */
-  private dbToProp: Record<string, string>
-  /** Built column schemas keyed by DB column name (for `cast_as`). */
-  private columnSchemas: Record<string, ColumnSchema>
-  /** Column builders keyed by BOTH property name and DB name. */
-  private v3Columns: Record<string, V3ColumnLike>
-  /** EQL 3.0.2+ requires query-domain casts PostgREST cannot express. */
-  private queryDomainsRequired: boolean
-
-  constructor(
-    tableName: string,
-    table: AnyV3Table,
-    encryptionClient: EncryptionClient,
-    supabaseClient: SupabaseClientLike,
-    allColumns: string[] | null = null,
-    queryDomainsRequired = false,
-  ) {
-    super(
-      tableName,
-      // The base class only ever calls BuildableTable members on the schema
-      // (build / encryptModel plumbing); every v2-specific behaviour is
-      // overridden below.
-      table as unknown as EncryptedTable<EncryptedTableColumn>,
-      encryptionClient,
-      supabaseClient,
-      allColumns,
-    )
-
-    this.v3Table = table
-    this.queryDomainsRequired = queryDomainsRequired
-    this.propToDb = table.buildColumnKeyMap()
-    this.columnSchemas = table.build().columns
-
-    this.dbToProp = Object.create(null) as Record<string, string>
-    for (const [property, dbName] of Object.entries(this.propToDb)) {
-      this.dbToProp[dbName] = property
-    }
-
-    assertNoPropertyDbNameCollision(tableName, this.propToDb, allColumns)
-
-    // Null-prototype: keyed by DB column names, and `validateTransforms` reads
-    // it without an own-key guard — an inherited `constructor`/`toString` would
-    // otherwise resolve truthy for a plaintext column of that name.
-    this.v3Columns = Object.create(null) as Record<string, V3ColumnLike>
-    for (const [property, builder] of Object.entries(table.columnBuilders)) {
-      if (builder instanceof EncryptedV3Column) {
-        const col = builder as unknown as V3ColumnLike
-        this.v3Columns[property] = col
-        this.v3Columns[col.getName()] = col
-      }
-    }
-
-    // The base class derives encrypted column names from build(), which v3
-    // keys by DB name. Filters and select strings address columns by JS
-    // property name, so recognition must cover both.
-    this.encryptedColumnNames = Object.keys(this.v3Columns)
-  }
-
-  // ---------------------------------------------------------------------------
-  // Dialect overrides
-  // ---------------------------------------------------------------------------
-
-  protected override getColumnMap(): Record<string, BuildableQueryColumn> {
-    return this.v3Columns as unknown as Record<string, BuildableQueryColumn>
-  }
-
-  /** Resolve a JS property name to its DB column name. `Object.hasOwn` guards
-   * the inherited-member hazard described on {@link EncryptedTable.buildColumnKeyMap}. */
-  private dbNameFor(name: string): string {
-    return Object.hasOwn(this.propToDb, name) ? this.propToDb[name] : name
-  }
-
-  protected override filterColumnName(column: string): DbName {
-    return this.dbNameFor(column) as DbName
-  }
-
-  /**
-   * `ORDER BY` on an OPE-backed column is supported; on every other encrypted
-   * column it is rejected.
-   *
-   * A bare `ORDER BY col` IS wrong. The `*_ord` domains are
-   * `CREATE DOMAIN … AS jsonb`, and the bundle declares no btree operator class
-   * on any domain — it actively lints against one (`domain_opclass`), because an
-   * opclass on a domain bypasses operator resolution. So the sort resolves
-   * through jsonb's default `jsonb_cmp` and compares the envelope's keys in
-   * storage order, starting at the random ciphertext `c`. No error, and a
-   * stable, meaningless row order. (Measured: over 10 rows it returns
-   * `r00,r04,r08,r01,…` where the plaintext order is `r00..r09`.)
-   *
-   * But the correct sort key is reachable without a function call. `eql_v3.ord_term`
-   * returns the domain's `op` term, and OPE is order-preserving by construction:
-   * ordering by the term reproduces the plaintext order. PostgREST cannot emit
-   * `ORDER BY eql_v3.ord_term(col)`, but it CAN emit a jsonb path —
-   * `order=col->op.asc` — which selects exactly that term. Measured against a
-   * live PostgREST: `order=amount->op.asc` and `.desc` both reproduce the
-   * plaintext order for `integer_ord` and `text_search`, over 10 rows.
-   *
-   * So the guard is on the ordering FLAVOUR, not on encryption:
-   *
-   * - `ope` present → order by `col->op`. Every plain `_ord` domain, plus
-   *   `text_ord` and `text_search`.
-   * - `ore` present → reject. The `ob` term is an array of ORE blocks whose
-   *   comparison needs the superuser-only opclass; a jsonb-path sort over it is
-   *   meaningless. (Such a column cannot hold data on managed Postgres at all:
-   *   its domain CHECK raises `ore_domain_unavailable`.)
-   * - neither → reject. Storage-only, equality-only and match-only columns
-   *   carry no ordering term to sort by.
-   *
-   * A column absent from {@link v3Columns} is a plaintext passthrough and orders
-   * normally. This runtime guard is the only protection the untyped
-   * (no-`schemas`) surface has.
-   */
-  protected override validateTransforms(): void {
-    for (const t of this.transforms) {
-      if (t.kind !== 'order') continue
-      const column = this.v3Columns[t.column]
-      if (!column) continue
-
-      const indexes = this.columnSchemas[column.getName()]?.indexes
-      if (indexes?.ope) continue
-
-      const reason = indexes?.ore
-        ? 'its ORE ordering term (`ob`) needs the superuser-only ORE operator class, which PostgREST cannot reach through a jsonb path'
-        : 'it carries no ordering term to sort by'
-
-      throw new Error(
-        `[supabase v3]: cannot order by encrypted column "${column.getName()}" (${column.getEqlType()}) — ${reason}. ` +
-          'Order by a plaintext column, or use an OPE-backed ordering domain ' +
-          '(`*_ord`, `text_ord`, `text_search`), or use the EQL v3 Drizzle integration.',
-      )
-    }
-  }
-
-  /**
-   * Encrypted ordering columns sort by their `op` term, not by the envelope.
-   *
-   * `order=col->op` is the one ordering expression PostgREST can emit that
-   * reaches the OPE term. It must NOT leak into filters — those compare whole
-   * envelopes through the `eql_v3.*` operators — which is why this is its own
-   * seam rather than a change to `filterColumnName`.
-   *
-   * The canonical EQL form is `ORDER BY eql_v3.ord_term(col)`, which returns
-   * `eql_v3_internal.ope_cllw` — a domain over `bytea`, ordered by the native
-   * btree. PostgREST cannot call a function, so it orders the `op` term where it
-   * sits, inside the envelope. The two agree because the term is what
-   * `ord_term()` returns.
-   *
-   * `->` (jsonb) rather than `->>` (text) keeps the comparison on the typed
-   * value. Note this does NOT avoid the database collation: Postgres compares
-   * jsonb strings with `varstr_cmp` under the default collation, exactly as it
-   * does text. What makes the ordering collation-independent is the term itself
-   * — fixed-width lowercase hex (`[0-9a-f]`, 130 chars for `integer_ord`, 82 for
-   * `text_search`) — and every collation orders digits before letters and hex
-   * letters among themselves. `match-bloom`'s sibling assertion pins that shape.
-   *
-   * This runs at column-name-mapping time (`transformToDbSpace`), BEFORE
-   * `buildAndExecuteQuery` calls `validateTransforms`. For an encrypted column
-   * with no `ope` index it therefore returns a bare `dbName` here — a name that
-   * would sort by `jsonb_cmp` over the ciphertext if it reached PostgREST — but
-   * it never does: `validateTransforms` throws (with a domain-specific reason)
-   * before the query executes, so the bare name is only ever an intermediate
-   * value on a request that is about to be rejected.
-   */
-  protected override orderColumnName(column: string): DbName {
-    const dbName = this.dbNameFor(column)
-    const encrypted = this.v3Columns[column]
-    if (!encrypted) return dbName as DbName
-
-    return (
-      this.columnSchemas[dbName]?.indexes?.ope ? `${dbName}->op` : dbName
-    ) as DbName
-  }
-
-  /**
-   * Resolve a raw `.filter()` operator to the capability it exercises. Unlike
-   * v2, a supported v3 operand is a full storage envelope, so `queryType`
-   * never selects a narrowing — it only tells {@link encryptCollectedTerms}
-   * which capability to demand of the column. Getting it wrong therefore
-   * produces a wrong accept/reject, not a wrong ciphertext: the base class's
-   * `'equality'` default rejects `.filter('bio', 'cs', …)` on a
-   * `public.eql_v3_text_match` column, the one query that column can answer.
-   *
-   * Unknown operators throw rather than silently defaulting to equality, which
-   * would encrypt a term the column may not even be able to compare.
-   */
-  protected override queryTypeForRawOp(operator: string): QueryTypeName {
-    switch (operator) {
-      case 'cs':
-        return 'freeTextSearch'
-      case 'gt':
-      case 'gte':
-      case 'lt':
-      case 'lte':
-        return 'orderAndRange'
-      case 'eq':
-      case 'neq':
-      case 'in':
-      case 'is':
-        return 'equality'
-      default:
-        throw new Error(
-          `[supabase v3]: unsupported raw filter operator "${operator}" on an encrypted column`,
-        )
-    }
-  }
-
-  protected override buildSelectString(): DbSelect | null {
-    if (this.selectColumns === null) return null
-    return addJsonbCastsV3(this.selectColumns, this.propToDb)
-  }
-
-  /**
-   * Expand the introspected column list (DB names) into JS property names.
-   *
-   * Load-bearing for `select('*')` on a DECLARED table that renames a column.
-   * `addJsonbCastsV3` only emits the `prop:db_name::jsonb` alias — the thing
-   * that makes PostgREST return the column under its property name — when the
-   * token it sees is a property name. Feeding it the raw DB name instead takes
-   * the unaliased `dbNames.has(...)` branch, so the row comes back keyed
-   * `created_at` while the declared row type promises `createdAt`, silently
-   * yielding `undefined` for a field TypeScript guarantees.
-   *
-   * A DB column with no encrypted builder (plaintext passthrough, and every
-   * synthesized column, where property == DB name) maps to itself.
-   */
-  protected override expandAllColumns(columns: string[]): string[] {
-    return columns.map((dbName) =>
-      Object.hasOwn(this.dbToProp, dbName) ? this.dbToProp[dbName] : dbName,
-    )
-  }
-
-  /** v3 domains are plain jsonb — send the raw payload, keyed by DB name. */
-  protected override transformEncryptedMutationModel(
-    model: Record<string, unknown>,
-  ): Record<string, unknown> {
-    const out: Record<string, unknown> = Object.create(null)
-    for (const [key, value] of Object.entries(model)) {
-      out[this.dbNameFor(key)] = value
-    }
-    return out
-  }
-
-  protected override transformEncryptedMutationModels(
-    models: Record<string, unknown>[],
-  ): Record<string, unknown>[] {
-    return models.map((model) => this.transformEncryptedMutationModel(model))
-  }
-
-  /**
-   * Validate a term's query type against its column's declared capabilities.
-   * Pure validation: `encrypt`/`bulkEncrypt` never receive the query type. On
-   * EQL 3.0.2+, free-text/JSON terms are rejected before this storage-encryption
-   * path can place ciphertext in a GET URL.
-   */
-  private assertTermQueryable(term: ScalarQueryTerm): V3ColumnLike {
-    const column = term.column as unknown as V3ColumnLike
-    let queryType = term.queryType ?? 'equality'
-    const capabilities = column.getQueryCapabilities()
-
-    // The `cs` wire operator is capability-overloaded: bloom free-text on a
-    // match/search TEXT column, encrypted ste_vec containment on a `types.Json`
-    // DOCUMENT column. Both arrive here as `freeTextSearch` (contains/matches/
-    // raw `cs` all map there); resolve to the capability the column actually
-    // carries. The two are mutually exclusive by construction, so this can
-    // never reinterpret a real free-text column.
-    if (
-      queryType === 'freeTextSearch' &&
-      !capabilities.freeTextSearch &&
-      capabilities.searchableJson
-    ) {
-      queryType = 'searchableJson'
-    }
-
-    if (
-      queryType !== 'equality' &&
-      queryType !== 'orderAndRange' &&
-      queryType !== 'freeTextSearch' &&
-      queryType !== 'searchableJson'
-    ) {
-      throw new Error(
-        `[supabase v3]: query type "${queryType}" is not supported on EQL v3 columns`,
-      )
-    }
-
-    if (!capabilities[queryType]) {
-      throw new Error(
-        `[supabase v3]: column "${column.getName()}" (${column.getEqlType()}) does not support ${queryType} queries — declare the column with a domain that carries that capability`,
-      )
-    }
-
-    if (queryType === 'freeTextSearch' || queryType === 'searchableJson') {
-      // This is the common boundary for every spelling that collects an
-      // encrypted match/containment term: matches(), contains(), not(), raw
-      // filter(), and both forms of or(). Method-level checks provide earlier
-      // errors for the direct helpers, but cannot cover the inherited raw
-      // filter paths on their own.
-      this.assertPostgrestCanQueryEncryptedOperator('filter', column.getName())
-    }
-
-    if (queryType === 'searchableJson') {
-      // THE single enforced operand boundary for encrypted-JSON containment.
-      // Terms reach this resolver from every spelling — contains(), raw
-      // .filter(col,'cs',…), not(col,'contains'|'matches',…), and .or()
-      // string/structured conditions — and only contains() has a method-level
-      // guard. Without this check a raw string (e.g. a free-text term ported
-      // from a text column, or an .or() condition value, which is always a
-      // string) would be storage-encrypted as a JSON SCALAR and silently match
-      // nothing; pre-#650 every such spelling failed loudly on capability.
-      assertJsonContainmentOperand(column.getName(), term.value)
-    }
-
-    // Free-text (bloom) needle floor. A needle shorter than the tokenizer's
-    // token_length produces NO tokens, so `bf @> '{}'` holds for every row and
-    // the query would silently return (and the caller decrypt) the whole table
-    // — a fail-open over-exposure. Reject it up front, mirroring the Drizzle v3
-    // adapter (matchNeedleError) so both first-party surfaces guard identically.
-    // JSON containment terms (searchableJson) are validated separately above.
-    if (queryType === 'freeTextSearch') {
-      const match = column.build().indexes?.match
-      const reason = match ? matchNeedleError(term.value, match) : undefined
-      if (reason) {
-        throw new Error(
-          `[supabase v3]: cannot search column "${column.getName()}": ${reason}`,
-        )
-      }
-    }
-
-    return column
-  }
-
-  private encryptionFailure(message: string, cause?: EncryptionError): never {
-    logger.error(
-      `Supabase: failed to encrypt query terms for table "${this.tableName}"`,
-    )
-    // Most callers pass the operation's own `EncryptionError`; the contract-
-    // violation cases (bulk length mismatch, null envelope) have none, so
-    // synthesize one — a broken query encryption is still an encryption failure,
-    // and callers branch on `error.encryptionError` regardless.
-    throw new EncryptionFailedError(
-      `Failed to encrypt query terms: ${message}`,
-      cause ?? { type: EncryptionErrorTypes.EncryptionError, message },
-    )
-  }
-
-  /**
-   * Encrypt every filter operand as a full storage envelope, serialized to jsonb
-   * text for the PostgREST filter value.
-   *
-   * Terms are grouped by column and each group takes ONE `bulkEncrypt` crossing.
-   * `in(col, [a, b, c])` collects one term per element (the list must never be
-   * encrypted whole), so encrypting per term spent N ZeroKMS/FFI round-trips
-   * where one would do. `bulkEncrypt` carries a single `{table, column}` for the
-   * whole payload, so the grouping is mandatory, not an optimisation: one bulk
-   * call over a mixed-column term array would stamp one column onto every
-   * plaintext. Results are scattered back onto the terms' original indices,
-   * which is the contract `termMap` downstream relies on.
-   *
-   * Mirrors `eql/v3/drizzle/operators.ts` `encryptOperands` — same batching
-   * contract, same length assertion, same fallback. Kept separate because that
-   * one encrypts a single-column operand list and returns `SQL[]`, while this
-   * must group a multi-column term array and preserve positions.
-   */
-  protected override async encryptCollectedTerms(
-    terms: ScalarQueryTerm[],
-  ): Promise<EncryptedQueryResult[]> {
-    const groups = new Map<
-      V3ColumnLike,
-      { indices: number[]; values: ScalarQueryTerm['value'][] }
-    >()
-    terms.forEach((term, index) => {
-      const column = this.assertTermQueryable(term)
-      const group = groups.get(column) ?? { indices: [], values: [] }
-      group.indices.push(index)
-      group.values.push(term.value)
-      groups.set(column, group)
-    })
-
-    const bulkEncrypt = this.encryptionClient.bulkEncrypt?.bind(
-      this.encryptionClient,
-    )
-    // Each term becomes the `JSON.stringify`'d storage envelope — a `string`,
-    // which is one arm of `EncryptedQueryResult`. PostgREST cannot cast a filter
-    // value to the `eql_v3.query_<name>` twins, so v3 sends full envelopes where
-    // v2 sends `encryptQuery` composite literals; both are `EncryptedQueryResult`.
-    const results = new Array<EncryptedQueryResult>(terms.length)
-
-    await Promise.all(
-      Array.from(groups, async ([column, { indices, values }]) => {
-        const encrypted = bulkEncrypt
-          ? await this.bulkEncryptGroup(bulkEncrypt, column, values)
-          : await this.encryptGroupPerTerm(column, values)
-
-        encrypted.forEach((envelope, i) => {
-          results[indices[i]] = JSON.stringify(envelope)
-        })
-      }),
-    )
-
-    return results
-  }
-
-  /** One FFI crossing for a column's whole operand list. */
-  private async bulkEncryptGroup(
-    bulkEncrypt: NonNullable<EncryptionClient['bulkEncrypt']>,
-    column: V3ColumnLike,
-    values: ScalarQueryTerm['value'][],
-  ): Promise<Array<Encrypted | null>> {
-    const baseOp = bulkEncrypt(
-      values.map((plaintext) => ({ plaintext })) as never,
-      { column, table: this.v3Table } as never,
-    )
-    const op = this.lockContext
-      ? baseOp.withLockContext(this.lockContext)
-      : baseOp
-    if (this.auditConfig) op.audit(this.auditConfig)
-
-    const result = await op
-    if (result.failure)
-      this.encryptionFailure(result.failure.message, result.failure)
-
-    // `bulkEncrypt` is position-stable, so a length mismatch means the contract
-    // was violated. Truncating instead would silently widen an `in` predicate
-    // (or narrow a `not.in`) to whatever came back. `result.data` is now
-    // `BulkEncryptedData` — `{ id?, data: Encrypted | null }[]` — not `unknown`.
-    const encrypted = result.data
-    if (encrypted.length !== values.length) {
-      this.encryptionFailure(
-        `bulk encryption returned ${encrypted.length} terms for ${values.length} values on column "${column.getName()}".`,
-      )
-    }
-    return encrypted.map((term, i) => {
-      // `BulkEncryptedData` types the element as `Encrypted | null`. A `null`
-      // envelope here would be `JSON.stringify`'d to the literal string `"null"`
-      // and sent as the filter operand — silently matching whatever `"null"`
-      // encodes to rather than failing. A query term should never encrypt to a
-      // null envelope, so treat it as a contract violation, not a value.
-      if (term.data === null) {
-        this.encryptionFailure(
-          `bulk encryption returned a null envelope at position ${i} for column "${column.getName()}".`,
-        )
-      }
-      return term.data
-    })
-  }
-
-  /** Fallback for a client that predates `bulkEncrypt`. */
-  private async encryptGroupPerTerm(
-    column: V3ColumnLike,
-    values: ScalarQueryTerm['value'][],
-  ): Promise<Encrypted[]> {
-    return Promise.all(
-      values.map(async (value) => {
-        const baseOp = this.encryptionClient.encrypt(value, {
-          column,
-          table: this.v3Table,
-        })
-        const op = this.lockContext
-          ? baseOp.withLockContext(this.lockContext)
-          : baseOp
-        if (this.auditConfig) op.audit(this.auditConfig)
-
-        const result = await op
-        if (result.failure) {
-          this.encryptionFailure(result.failure.message, result.failure)
-        }
-        return result.data
-      }),
-    )
-  }
-
-  /** Warn once per (op, column) that a `like`/`ilike` was delegated to `matches`. */
-  private static readonly warnedLikeDelegation = new Set<string>()
-
-  /** True when `column` is one of this table's encrypted v3 columns. */
-  private isEncryptedV3Column(column: string): boolean {
-    return Boolean(this.v3Columns[column])
-  }
-
-  /** True when `column` is an encrypted `types.Json` document column. */
-  private isSearchableJsonColumn(column: string): boolean {
-    const builder: V3ColumnLike | undefined = this.v3Columns[column]
-    return Boolean(builder?.getQueryCapabilities().searchableJson)
-  }
-
-  private assertPostgrestCanQueryEncryptedOperator(
-    method: string,
-    column: string,
-  ): void {
-    if (!this.queryDomainsRequired) return
-    throw new Error(
-      `[supabase v3]: ${method}() on encrypted column "${column}" is unavailable with EQL 3.0.2+: the SQL operator requires an eql_v3.query_* cast that PostgREST cannot express. Use the Drizzle or Prisma Next adapter, or a scoped SQL/RPC function.`,
-    )
-  }
-
-  /**
-   * `contains` on the v3 surface is EXACT containment: native jsonb/array `@>`
-   * on a plaintext column, ENCRYPTED ste_vec `@>` on a `types.Json` column (the
-   * sub-document operand is storage-encrypted whole; every leaf must match at
-   * its path — #650). On an encrypted match/search TEXT column containment is
-   * not the operation (that is the fuzzy `matches`), so refuse loudly rather
-   * than silently emit a bloom match under a name that promises exactness.
-   */
-  override contains(column: string, value: unknown): this {
-    if (this.isSearchableJsonColumn(column)) {
-      this.assertPostgrestCanQueryEncryptedOperator('contains', column)
-      // Same validator the term resolver enforces — failing here just surfaces
-      // the error at the call site instead of at execution.
-      assertJsonContainmentOperand(column, value)
-      return super.contains(column, value)
-    }
-    if (this.isEncryptedV3Column(column)) {
-      throw new Error(
-        `[supabase v3]: contains() is native (exact) containment and does not apply to encrypted column "${column}". Use matches() for encrypted free-text search.`,
-      )
-    }
-    return super.contains(column, value)
-  }
-
-  /**
-   * `matches` is the encrypted free-text operator: fuzzy bloom-filter token
-   * matching, one-sided (may false-positive), NOT containment. It requires an
-   * encrypted match/search column; on a plaintext column, `contains` (native
-   * `@>`) is what the caller means — and on an encrypted JSON column,
-   * `contains`/`selectorEq` are (matching a document is containment, not
-   * free-text). Guarded here because both spellings collect the same
-   * `freeTextSearch` term, which the capability resolver would otherwise
-   * silently accept as containment of the raw string.
-   */
-  override matches(column: string, value: unknown): this {
-    if (this.isSearchableJsonColumn(column)) {
-      throw new Error(
-        `[supabase v3]: matches() is encrypted free-text search and does not apply to encrypted JSON column "${column}". Use contains("${column}", subDocument) or selectorEq("${column}", path, value).`,
-      )
-    }
-    if (!this.isEncryptedV3Column(column)) {
-      throw new Error(
-        `[supabase v3]: matches() is encrypted free-text search and requires an encrypted column; "${column}" is not one. Use contains() for native containment.`,
-      )
-    }
-    this.assertPostgrestCanQueryEncryptedOperator('matches', column)
-    return super.matches(column, value)
-  }
-
-  /**
-   * `not(col, 'contains', …)` on an encrypted TEXT column would negate a fuzzy
-   * bloom match under the `contains` name — the exact confusion #617 removes —
-   * because the base `not()` path rewrites the `contains` spelling to the `cs`
-   * wire operator. Reject it and steer to the `matches` spelling (or the raw
-   * `cs` operator, which is honest about the wire op).
-   *
-   * On an encrypted JSON column negated containment IS the honest exact
-   * operation (`not.cs` over ste_vec containment — {@link selectorNe} compiles
-   * to it), so it passes through. Plaintext columns keep native negated
-   * containment, and every other operator is delegated unchanged.
-   */
-  override not(column: string, operator: string, value: unknown): this {
-    if (
-      operator === 'contains' &&
-      this.isEncryptedV3Column(column) &&
-      !this.isSearchableJsonColumn(column)
-    ) {
-      throw new Error(
-        `[supabase v3]: not("${column}", 'contains', …) does not apply to encrypted column "${column}" — that is fuzzy free-text matching, not containment. Use not("${column}", 'matches', …) or the raw 'cs' operator.`,
-      )
-    }
-    // Mirror of the matches() guard: a `matches` spelling on a JSON column
-    // would otherwise resolve to containment (the two share the `cs` wire op),
-    // silently negating an EXACT operation under a name that promises FUZZY.
-    if (operator === 'matches' && this.isSearchableJsonColumn(column)) {
-      throw new Error(
-        `[supabase v3]: not("${column}", 'matches', …) does not apply to encrypted JSON column "${column}" — matches() is free-text search. Use not("${column}", 'contains', subDocument) or selectorNe("${column}", path, value).`,
-      )
-    }
-    return super.not(column, operator, value)
-  }
-
-  /**
-   * Validate + reconstruct a selector needle: `('$.user.role', 'admin')` →
-   * `{user: {role: 'admin'}}`. Shared by {@link selectorEq}/{@link selectorNe};
-   * throws with column context for a non-JSON column, an invalid path, or a
-   * non-scalar leaf.
-   */
-  private selectorNeedle(
-    method: string,
-    column: string,
-    path: string,
-    value: unknown,
-  ): Record<string, unknown> {
-    if (!this.isSearchableJsonColumn(column)) {
-      throw new Error(
-        `[supabase v3]: ${method}() requires an encrypted JSON (types.Json) column; "${column}" is not one.`,
-      )
-    }
-    // Selector comparisons compare a scalar LEAF (null included in the shared
-    // helper's rejection; eq/ne arm — `ordering: false`;
-    // PostgREST cannot express selector ordering yet, see
-    // cipherstash/encrypt-query-language#407).
-    const leafReason = unsupportedLeafReason(value, false)
-    if (leafReason) {
-      throw new Error(
-        `[supabase v3]: ${method}("${column}", "${path}", …): ${leafReason}`,
-      )
-    }
-    // Stricter than the shared helper (whose Date/bigint arms serve the Drizzle
-    // surface): a stored JsonDocument leaf is a JSON scalar, so a Date/bigint
-    // needle could never match one — reject with the serialization steer
-    // instead of running a query that structurally returns nothing.
-    if (
-      typeof value !== 'string' &&
-      typeof value !== 'number' &&
-      typeof value !== 'boolean'
-    ) {
-      throw new Error(
-        `[supabase v3]: ${method}("${column}", "${path}", …): a JSON document leaf is a JSON scalar (string/number/boolean); got ${value instanceof Date ? 'a Date — pass date.toISOString() (or the stored form)' : typeof value}.`,
-      )
-    }
-    let segments: string[]
-    try {
-      segments = parseSelectorSegments(path)
-    } catch (err) {
-      throw new Error(
-        `[supabase v3]: ${method}("${column}", …): ${err instanceof Error ? err.message : String(err)}`,
-      )
-    }
-    return reconstructSelectorDocument(segments, value)
-  }
-
-  /**
-   * Encrypted JSONPath-selector equality: matches rows whose document carries
-   * exactly `value` at `path`. Equality at a path IS containment of the
-   * path-shaped needle (`{user: {role: 'admin'}}`), so this compiles to
-   * {@link contains} — the ste_vec entry at the selector matches on its
-   * equality/ordering term. Selector ORDERING (`gt`/`lt`/…) is not expressible
-   * over PostgREST until the bundle grows a needle-comparison overload
-   * (cipherstash/encrypt-query-language#407); the Drizzle adapter's
-   * `ops.selector()` supports it today.
-   */
-  selectorEq(column: string, path: string, value: unknown): this {
-    this.assertPostgrestCanQueryEncryptedOperator('selectorEq', column)
-    const needle = this.selectorNeedle('selectorEq', column, path, value)
-    return super.contains(column, needle)
-  }
-
-  /**
-   * Encrypted JSONPath-selector inequality: rows whose document does NOT carry
-   * `value` at `path` — INCLUDING rows where the path is absent AND rows whose
-   * document column is SQL NULL, matching the Drizzle selector's `ne` (whose
-   * `OR entry IS NULL` arm covers both absence cases). A bare `not.cs` would
-   * drop NULL documents under three-valued logic (`NOT (NULL @> x)` is NULL),
-   * so this compiles to a structured OR:
-   * `column.is.null, column.not.cs.<needle>` — the containment condition's
-   * operand is encrypted through the normal or-condition term path.
-   */
-  selectorNe(column: string, path: string, value: unknown): this {
-    this.assertPostgrestCanQueryEncryptedOperator('selectorNe', column)
-    const needle = this.selectorNeedle('selectorNe', column, path, value)
-    return super.or([
-      { column, op: 'is', value: null },
-      { column, op: 'contains', negate: true, value: needle },
-    ])
-  }
-
-  /**
-   * `like`/`ilike` on an ENCRYPTED column are a best-effort compatibility shim,
-   * delegated to `matches`. EQL v3 free-text search is fuzzy bloom token
-   * matching, not SQL pattern matching, so the result is APPROXIMATE — matching
-   * is case-insensitive and one-sided (may false-positive), and anchoring is
-   * lost. Leading/trailing `%` are stripped; an internal `%` or any `_` cannot be
-   * approximated by trigram matching and throws. A plaintext column keeps real
-   * SQL LIKE.
-   */
-  override like(column: string, pattern: string): this {
-    if (!this.isEncryptedV3Column(column)) return super.like(column, pattern)
-    return this.matches(column, this.likeNeedle(column, 'like', pattern))
-  }
-
-  override ilike(column: string, pattern: string): this {
-    if (!this.isEncryptedV3Column(column)) return super.ilike(column, pattern)
-    return this.matches(column, this.likeNeedle(column, 'ilike', pattern))
-  }
-
-  /**
-   * Reduce a SQL LIKE pattern to a fuzzy-match needle, or throw when it cannot be
-   * approximated. Strips surrounding `%` (prefix/suffix wildcards, which fuzzy
-   * matching subsumes); an internal `%` or any `_` is unapproximable. Warns once
-   * per (op, column) that the delegation is approximate.
-   */
-  private likeNeedle(column: string, op: string, pattern: string): string {
-    const needle = pattern.replace(/^%+/, '').replace(/%+$/, '')
-    if (needle.includes('%') || pattern.includes('_')) {
-      throw new Error(
-        `[supabase v3]: "${op}" pattern "${pattern}" on encrypted column "${column}" has wildcards fuzzy free-text matching cannot honor (an internal "%" or any "_"). Use matches("${column}", term) with a literal search term.`,
-      )
-    }
-    const key = `${op}:${column}`
-    if (!EncryptedQueryBuilderV3Impl.warnedLikeDelegation.has(key)) {
-      EncryptedQueryBuilderV3Impl.warnedLikeDelegation.add(key)
-      logger.warn(
-        `[supabase v3]: "${op}" on encrypted column "${column}" is delegated to matches() (fuzzy bloom token search). Results are APPROXIMATE — case-insensitive, one-sided (may false-positive), and wildcards/anchoring are not honored. Call matches() directly to make this explicit.`,
-      )
-    }
-    return needle
-  }
-
-  /**
-   * Encrypted `matches` goes through the bloom-filter `@>`, which the bundle
-   * declares on the domain as PostgREST's `cs`. The operand is the full storage
-   * envelope; `eql_v3.matches` (the SQL function) extracts the `bf` array from
-   * both sides.
-   *
-   * Emitted via `filter(col, 'cs', json)` rather than `q.contains(col, json)`:
-   * postgrest-js's `contains` re-serializes a non-string operand, and our
-   * operand is already `JSON.stringify`d. Plaintext `contains` (not encrypted)
-   * falls through to the base's native path.
-   */
-  protected override applyContainsFilter(
-    q: SupabaseQueryBuilder,
-    column: DbName,
-    value: unknown,
-    wasEncrypted: boolean,
-  ): SupabaseQueryBuilder {
-    if (wasEncrypted) {
-      this.assertPostgrestCanQueryEncryptedOperator('filter', column)
-      return q.filter(column, 'cs', value)
-    }
-    return super.applyContainsFilter(q, column, value, wasEncrypted)
-  }
-
-  /**
-   * `.or()` string conditions carry raw PostgREST operators, so a free-text
-   * condition arrives as `cs` — not a {@link FilterOp}. Resolve it through the
-   * same table the raw `.filter()` path uses, so `.or('amount.cs.5')` on an
-   * `integer_ord` column is rejected by the capability guard rather than
-   * silently encrypted as an equality term. A structured `{ op: 'matches' }`
-   * condition maps to free-text directly.
-   */
-  protected override queryTypeForOrOp(op: FilterOp): QueryTypeName {
-    if (op === 'matches') return 'freeTextSearch'
-    // Structured conditions may carry the `contains` METHOD spelling (the wire
-    // token becomes `cs` in rebuildOrString). It maps to the same capability
-    // gate as `cs`; on a JSON column the term resolver then re-types it to
-    // searchableJson and validates the operand. selectorNe's IS-NULL-inclusive
-    // or-form relies on this arm.
-    if (op === 'contains') return 'freeTextSearch'
-    return this.queryTypeForRawOp(op)
-  }
-
-  /** Rebuild `Date` values from the encrypt-config `cast_as` (date/timestamp),
-   * mirroring the typed v3 client's decrypt-model path. */
-  protected override postprocessDecryptedRow(
-    row: Record<string, unknown>,
-  ): Record<string, unknown> {
-    // Every key an encrypted column can appear under: the keys this select
-    // actually produces (including caller-chosen aliases like `ts:createdAt`),
-    // plus the static property and DB names as a fallback for paths that record
-    // no select. Aliases win. Derived here from `this.selectColumns` (the row in
-    // hand) rather than cached from `buildSelectString`, so a reused builder can
-    // never postprocess a row with a previous operation's stale select map.
-    const keyToDb: Record<string, string> = Object.assign(
-      Object.create(null),
-      this.selectColumns === null
-        ? undefined
-        : selectKeyToDbV3(this.selectColumns, this.propToDb),
-    )
-    for (const [property, dbName] of Object.entries(this.propToDb)) {
-      keyToDb[property] ??= dbName
-      keyToDb[dbName] ??= dbName
-    }
-
-    const out: Record<string, unknown> = { ...row }
-    for (const [key, dbName] of Object.entries(keyToDb)) {
-      const castAs = this.columnSchemas[dbName]?.cast_as
-      if (!DATE_LIKE_CAST_SET.has(castAs as string)) continue
-      const value = out[key]
-      if (value == null || value instanceof Date) continue
-      if (typeof value === 'string' || typeof value === 'number') {
-        out[key] = new Date(value)
-      }
-    }
-    return out
-  }
-}
